+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fixed not all UI elements updating immediately when you changed the deck tracker's language.
 
 **Battlegrounds**:
+- Added counters for Eternal Knight and Ancestral Automaton.
 - Fixed the session widget not appearing immediately when starting the deck tracker while mid-match.
 - Fixed a 1 pixel gap above overlay elements at the top of the screen.
 - Fixed a recent issue with incorrect combat odds around Polarizing Beatboxer.